Clarify the need for submodule cloning
Issue #21 highlights a potential issue, in that it is not made clear
that you need to pull down the submodules before building. This updates
the README to make this a little more clear.

Also, seems the `setup.sh` script had a potential bug, where the syntax
used is not supported by all versions of sh (double brackets `[[`). So
the default executable is changed to bash, and for good measure the
submodule stuff is done first, just in case.